+++ src/AjgarlagPsrHttpMessageBundle.php
@@ -14,6 +14,9 @@ namespace Ajgarlag\Bundle\PsrHttpMessageBundle;
 
 use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\RegisterHttpMessageFactoriesPass;
 use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\RegisterNyholmPsr17FactoriesPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\TagArgumentValueResolverPass;
+use Ajgarlag\Bundle\PsrHttpMessageBundle\DependencyInjection\Compiler\TagViewEventListenerPass;
+use Symfony\Component\DependencyInjection\Compiler\PassConfig;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
@@ -23,7 +26,9 @@ class AjgarlagPsrHttpMessageBundle extends Bundle
     {
         parent::build($container);
 
+        $container->addCompilerPass(new TagArgumentValueResolverPass(), PassConfig::TYPE_BEFORE_OPTIMIZATION, 1);
         $container->addCompilerPass(new RegisterHttpMessageFactoriesPass());
         $container->addCompilerPass(new RegisterNyholmPsr17FactoriesPass());
+        $container->addCompilerPass(new TagViewEventListenerPass());
     }
 }